refactor : CloseButton -> CloseXButton 으로 이름 수정
