+++ modal-component/src/Modal/CloseXButton.tsx
