+++ payments-hook/src/App.tsx
@@ -1,11 +1,6 @@
 import './App.css';
-// import { useCardNumbers } from 'pongda-payments-hooks';
-import useCardNumbers from './useCardNumbers/useCardNumbers';
-import getCardNetwork from './useCardNumbers/getCardNetwork';
 import useFormattedCardNumbers from './useFormattedCardNumbers/useFormattedCardNumbers';
 function App() {
-  // const { cardNumbers, handleCardNumbersChange, isError, errorMessage } =
-  //   useCardNumbers();
   const {
     formattedCardNumbers,
     handleFormattedCardNumbersChange,
@@ -21,22 +16,6 @@ function App() {
         onChange={handleFormattedCardNumbersChange}
       ></input>
 
-      {/* <input
-        value={cardNumbers.FIRST}
-        onChange={handleCardNumbersChange({ target: 'FIRST' })}
-      ></input>
-      <input
-        value={cardNumbers.SECOND}
-        onChange={handleCardNumbersChange({ target: 'SECOND' })}
-      ></input>
-      <input
-        value={cardNumbers.THIRD}
-        onChange={handleCardNumbersChange({ target: 'THIRD' })}
-      ></input>
-      <input
-        value={cardNumbers.FOURTH}
-        onChange={handleCardNumbersChange({ target: 'FOURTH' })}
-      ></input> */}
       {isError && (
         <div className="error-message">
           {Object.values(errorMessage).map((message, index) => (
